+++ src/IRProtocol.hpp
@@ -63,8 +63,8 @@ const char string_Kaseikyo_Mitsubishi[] PROGMEM = "Kaseikyo_Mitsubishi";
 const char string_RC5[] PROGMEM = "RC5";
 const char string_RC6[] PROGMEM = "RC6";
 const char string_Samsung[] PROGMEM = "Samsung";
-const char string_Samsung48[] PROGMEM = "Samsung48";
 const char string_SamsungLG[] PROGMEM = "SamsungLG";
+const char string_Samsung48[] PROGMEM = "Samsung48";
 const char string_Sharp[] PROGMEM = "Sharp";
 const char string_Sony[] PROGMEM = "Sony";
 const char string_BangOlufsen[] PROGMEM = "Bang&Olufsen";
@@ -80,7 +80,7 @@ const char string_FAST[] PROGMEM = "FAST";
 const char *const ProtocolNames[]
 PROGMEM = { string_Unknown, string_PulseWidth, string_PulseDistance, string_Apple, string_Denon, string_JVC, string_LG, string_LG2,
         string_NEC, string_NEC2, string_Onkyo, string_Panasonic, string_Kaseikyo, string_Kaseikyo_Denon, string_Kaseikyo_Sharp,
-        string_Kaseikyo_JVC, string_Kaseikyo_Mitsubishi, string_RC5, string_RC6, string_Samsung, string_Samsung48, string_SamsungLG,
+        string_Kaseikyo_JVC, string_Kaseikyo_Mitsubishi, string_RC5, string_RC6, string_Samsung, string_SamsungLG, string_Samsung48,
         string_Sharp, string_Sony
 #if !defined(EXCLUDE_EXOTIC_PROTOCOLS)
         , string_BangOlufsen, string_BoseWave, string_Lego, string_MagiQuest, string_Whynter, string_FAST
@@ -178,7 +178,7 @@ void printIRResultShort(Print *aSerial, IRData *aIRDataPtr, bool aPrintRepeatGap
 #endif
 #if !defined(DISABLE_CODE_FOR_RECEIVER)
         aSerial->print(' ');
-        aSerial->print((aIRDataPtr->rawDataPtr->rawlen + 1) / 2, DEC);
+        aSerial->print((aIRDataPtr->rawlen + 1) / 2, DEC);
         aSerial->println(F(" bits (incl. gap and start) received"));
 #endif
     } else {
@@ -218,7 +218,7 @@ void printIRResultShort(Print *aSerial, IRData *aIRDataPtr, bool aPrintRepeatGap
 #if !defined(DISABLE_CODE_FOR_RECEIVER)
             if (aPrintRepeatGap) {
                 aSerial->print(F(" gap="));
-                aSerial->print((uint32_t) aIRDataPtr->rawDataPtr->rawbuf[0] * MICROS_PER_TICK);
+                aSerial->print((uint32_t) aIRDataPtr->initialGap * MICROS_PER_TICK);
                 aSerial->print(F("us"));
             }
 #else