+++ src/IRReceive.hpp
@@ -342,6 +342,19 @@ void IRrecv::start() {
     pinModeFast(_IR_TIMING_TEST_PIN, OUTPUT);
 #endif
 }
+
+/*
+ * Do not resume() reading of IR data
+ */
+void IRrecv::restartTimer() {
+    // Setup for cyclic 50 us interrupt
+    timerConfigForReceive(); // no interrupts enabled here!
+    // Timer interrupt is enabled after state machine reset
+    timerEnableReceiveInterrupt(); // Enables the receive sample timer interrupt which consumes a small amount of CPU every 50 us.
+#ifdef _IR_MEASURE_TIMING
+    pinModeFast(_IR_TIMING_TEST_PIN, OUTPUT);
+#endif
+}
 /**
  * Alias for start().
  */
@@ -358,9 +371,17 @@ void IRrecv::start(uint32_t aMicrosecondsToAddToGapCounter) {
     irparams.TickCounterForISR += aMicrosecondsToAddToGapCounter / MICROS_PER_TICK;
     start();
 }
+void IRrecv::restartTimer(uint32_t aMicrosecondsToAddToGapCounter) {
+    irparams.TickCounterForISR += aMicrosecondsToAddToGapCounter / MICROS_PER_TICK;
+    restartTimer();
+}
 void IRrecv::startWithTicksToAdd(uint16_t aTicksToAddToGapCounter) {
     irparams.TickCounterForISR += aTicksToAddToGapCounter;
     start();
+}
+void IRrecv::restartTimerWithTicksToAdd(uint16_t aTicksToAddToGapCounter) {
+    irparams.TickCounterForISR += aTicksToAddToGapCounter;
+    restartTimer();
 }
 
 void IRrecv::addTicksToInternalTickCounter(uint16_t aTicksToAddToInternalTickCounter) {
@@ -383,6 +404,10 @@ void IRrecv::restartAfterSend() {
  * Disables the timer for IR reception.
  */
 void IRrecv::stop() {
+    timerDisableReceiveInterrupt();
+}
+
+void IRrecv::stopTimer() {
     timerDisableReceiveInterrupt();
 }
 /**
@@ -408,7 +433,7 @@ bool IRrecv::isIdle() {
 
 /**
  * Restart the ISR (Interrupt Service Routine) state machine, to enable receiving of the next IR frame.
- * Counting of gap timing is independent of StateForISR and therefore independent of call time of resume().
+ * Internal counting of gap timing is independent of StateForISR and therefore independent of call time of resume().
  */
 void IRrecv::resume() {
     // This check allows to call resume at arbitrary places or more than once
@@ -434,11 +459,16 @@ void IRrecv::initDecodedIRData() {
     } else {
         decodedIRData.flags = IRDATA_FLAGS_EMPTY;
         // save last protocol, command and address for repeat handling (where they are compared or copied back :-))
-        lastDecodedProtocol = decodedIRData.protocol; // repeat patterns can be equal between protocols (e.g. NEC and LG), so we must keep the original one
+        lastDecodedProtocol = decodedIRData.protocol; // repeat patterns can be equal between protocols (e.g. NEC, Samsung and LG), so we must keep the original one
         lastDecodedCommand = decodedIRData.command;
         lastDecodedAddress = decodedIRData.address;
 
     }
+
+    //These 2 variables allow to call resume() directly after decode, if no dump is required. since 4.3.0.
+    decodedIRData.initialGap = decodedIRData.rawDataPtr->rawbuf[0];
+    decodedIRData.rawlen = decodedIRData.rawDataPtr->rawlen;
+
     decodedIRData.protocol = UNKNOWN;
     decodedIRData.command = 0;
     decodedIRData.address = 0;
@@ -834,7 +864,7 @@ void IRrecv::initBiphaselevel(uint_fast8_t aRCDecodeRawbuffOffset, uint16_t aBip
 uint_fast8_t IRrecv::getBiphaselevel() {
     uint_fast8_t tLevelOfCurrentInterval; // 0 (SPACE) or 1 (MARK)
 
-    if (sBiphaseDecodeRawbuffOffset >= decodedIRData.rawDataPtr->rawlen) {
+    if (sBiphaseDecodeRawbuffOffset >= decodedIRData.rawlen) {
         return SPACE;  // After end of recorded buffer, assume space.
     }
 
@@ -917,7 +947,7 @@ bool IRrecv::decodeHash() {
     unsigned long hash = FNV_BASIS_32; // the result is the same no matter if we use a long or unsigned long variable
 
 // Require at least 6 samples to prevent triggering on noise
-    if (decodedIRData.rawDataPtr->rawlen < 6) {
+    if (decodedIRData.rawlen < 6) {
         return false;
     }
 #if RAW_BUFFER_LENGTH <= 254        // saves around 75 bytes program memory and speeds up ISR
@@ -925,7 +955,7 @@ bool IRrecv::decodeHash() {
 #else
     unsigned int i;
 #endif
-    for (i = 1; (i + 2) < decodedIRData.rawDataPtr->rawlen; i++) {
+    for (i = 1; (i + 2) < decodedIRData.rawlen; i++) {
         uint_fast8_t value = compare(decodedIRData.rawDataPtr->rawbuf[i], decodedIRData.rawDataPtr->rawbuf[i + 2]);
         // Add value into the hash
         hash = (hash * FNV_PRIME_32) ^ value;
@@ -1114,10 +1144,10 @@ bool IRrecv::checkForRecordGapsMicros(Print *aSerial) {
      * is smaller than known value for protocols (Sony with around 24 ms)
      */
     if (decodedIRData.protocol <= PULSE_DISTANCE
-            && decodedIRData.rawDataPtr->rawbuf[0] < (RECORD_GAP_MICROS_WARNING_THRESHOLD / MICROS_PER_TICK)) {
+            && decodedIRData.initialGap < (RECORD_GAP_MICROS_WARNING_THRESHOLD / MICROS_PER_TICK)) {
         aSerial->println();
         aSerial->print(F("Space of "));
-        aSerial->print(decodedIRData.rawDataPtr->rawbuf[0] * MICROS_PER_TICK);
+        aSerial->print(decodedIRData.initialGap * MICROS_PER_TICK);
         aSerial->print(F(" us between two detected transmission is smaller than the minimal gap of "));
         aSerial->print(RECORD_GAP_MICROS_WARNING_THRESHOLD);
         aSerial->println(F(" us known for implemented protocols like NEC, Sony, RC% etc.."));
@@ -1234,7 +1264,7 @@ void IRrecv::printDistanceWidthTimingInfo(Print *aSerial, DistanceWidthTimingInf
 
 uint32_t IRrecv::getTotalDurationOfRawData() {
     uint16_t tSumOfDurationTicks = 0;
-    for (uint_fast8_t i = 1; i < decodedIRData.rawDataPtr->rawlen; i++) {
+    for (uint_fast8_t i = 1; i < decodedIRData.rawlen; i++) {
         tSumOfDurationTicks += decodedIRData.rawDataPtr->rawbuf[i];
     }
     return tSumOfDurationTicks * (uint32_t) MICROS_PER_TICK;
@@ -1248,19 +1278,27 @@ uint32_t IRrecv::getTotalDurationOfRawData() {
  * @param aSerial The Print object on which to write, for Arduino you can use &Serial.
  */
 void IRrecv::printIRSendUsage(Print *aSerial) {
-    if (decodedIRData.protocol != UNKNOWN
-            && (decodedIRData.flags & (IRDATA_FLAGS_IS_AUTO_REPEAT | IRDATA_FLAGS_IS_REPEAT)) == 0x00) {
+    if (decodedIRData.flags & IRDATA_FLAGS_IS_PROTOCOL_WITH_DIFFERENT_REPEAT) {
+        /*
+         * Here we have a repeat of type NEC2 or SamsungLG. -> Inform the user to use this and not the initial protocol for sending.
+         */
+        Serial.print(F("!Use the "));
+        Serial.print(getProtocolString());
+        Serial.println(F(" protocol for sending!"));
+    } else {
+        if (decodedIRData.protocol != UNKNOWN
+                && (decodedIRData.flags & (IRDATA_FLAGS_IS_AUTO_REPEAT | IRDATA_FLAGS_IS_REPEAT)) == 0x00) {
 #if defined(DECODE_DISTANCE_WIDTH)
         uint_fast8_t tNumberOfArrayData = 0;
         if (decodedIRData.protocol == PULSE_DISTANCE || decodedIRData.protocol == PULSE_WIDTH) {
 #  if __INT_WIDTH__ < 32
-            aSerial->print(F("Send on a 8 bit platform with:"));
+            aSerial->print(F("Send on a 8 bit platform with: "));
             tNumberOfArrayData = ((decodedIRData.numberOfBits - 1) / 32) + 1;
             if(tNumberOfArrayData > 1) {
                 aSerial->println();
                 aSerial->print(F("    uint32_t tRawData[]={0x"));
 #  else
-                aSerial->print(F("Send on a 32 bit platform with:"));
+                aSerial->print(F("Send on a 32 bit platform with: "));
             tNumberOfArrayData = ((decodedIRData.numberOfBits - 1) / 64) + 1;
             if(tNumberOfArrayData > 1) {
                 aSerial->println();
@@ -1277,19 +1315,22 @@ void IRrecv::printIRSendUsage(Print *aSerial) {
                     }
                 }
                 aSerial->println(F("};"));
-                aSerial->print(F("   "));
+                aSerial->print(F("    "));
             }
+        } else {
+            aSerial->print(F("Send with: "));
         }
-        aSerial->print(F(" IrSender.send"));
+        aSerial->print(F("IrSender.send"));
+
 #else
-        aSerial->print(F("Send with: IrSender.send"));
+            aSerial->print(F("Send with: IrSender.send"));
 #endif
 
 #if defined(DECODE_DISTANCE_WIDTH)
         if (decodedIRData.protocol != PULSE_DISTANCE && decodedIRData.protocol != PULSE_WIDTH) {
 #endif
-        aSerial->print(getProtocolString());
-        aSerial->print(F("(0x"));
+            aSerial->print(getProtocolString());
+            aSerial->print(F("(0x"));
 #if defined(DECODE_MAGIQUEST)
             if (decodedIRData.protocol == MAGIQUEST) {
 #  if (__INT_WIDTH__ < 32)
@@ -1301,20 +1342,20 @@ void IRrecv::printIRSendUsage(Print *aSerial) {
                 aSerial->print(decodedIRData.address, HEX);
             }
 #else
-        /*
-         * New decoders have address and command
-         */
-        aSerial->print(decodedIRData.address, HEX);
+            /*
+             * New decoders have address and command
+             */
+            aSerial->print(decodedIRData.address, HEX);
 #endif
 
-        aSerial->print(F(", 0x"));
-        aSerial->print(decodedIRData.command, HEX);
-        if (decodedIRData.protocol == SONY) {
-            aSerial->print(F(", 2, "));
-            aSerial->print(decodedIRData.numberOfBits);
-        } else {
-            aSerial->print(F(", <numberOfRepeats>"));
-        }
+            aSerial->print(F(", 0x"));
+            aSerial->print(decodedIRData.command, HEX);
+            if (decodedIRData.protocol == SONY) {
+                aSerial->print(F(", 2, "));
+                aSerial->print(decodedIRData.numberOfBits);
+            } else {
+                aSerial->print(F(", <numberOfRepeats>"));
+            }
 
 #if defined(DECODE_DISTANCE_WIDTH)
         } else {
@@ -1357,14 +1398,10 @@ void IRrecv::printIRSendUsage(Print *aSerial) {
             aSerial->print(decodedIRData.extra, HEX);
         }
 #endif
-        aSerial->print(F(");"));
-        aSerial->println();
+            aSerial->print(F(");"));
+            aSerial->println();
+        }
     }
-#if defined(DECODE_DISTANCE_WIDTH)
-    else {
-        aSerial->print(F("Send with:"));
-    }
-#endif
 }
 
 /**
@@ -1386,7 +1423,7 @@ void IRrecv::printIRResultMinimal(Print *aSerial) {
 #  endif
 #endif
         aSerial->print(' ');
-        aSerial->print((decodedIRData.rawDataPtr->rawlen + 1) / 2, DEC);
+        aSerial->print((decodedIRData.rawlen + 1) / 2, DEC);
         aSerial->println(F(" bits received"));
     } else {
         /*
@@ -1419,7 +1456,7 @@ void IRrecv::printIRResultMinimal(Print *aSerial) {
  */
 void IRrecv::printIRResultRawFormatted(Print *aSerial, bool aOutputMicrosecondsInsteadOfTicks) {
 
-    uint8_t tRawlen = decodedIRData.rawDataPtr->rawlen; // Get it once here in order to print quite consistent data, even if ISR is running
+    uint8_t tRawlen = decodedIRData.rawlen; // Get it once here in order to print quite consistent data, even if ISR is running
 // Print Raw data
     aSerial->print(F("rawData["));
     aSerial->print(tRawlen, DEC);
@@ -1430,9 +1467,9 @@ void IRrecv::printIRResultRawFormatted(Print *aSerial, bool aOutputMicrosecondsI
      */
     aSerial->print(F(" -"));
     if (aOutputMicrosecondsInsteadOfTicks) {
-        aSerial->println((uint32_t) decodedIRData.rawDataPtr->rawbuf[0] * MICROS_PER_TICK, DEC);
+        aSerial->println((uint32_t) decodedIRData.initialGap * MICROS_PER_TICK, DEC);
     } else {
-        aSerial->println(decodedIRData.rawDataPtr->rawbuf[0], DEC);
+        aSerial->println(decodedIRData.initialGap, DEC);
     }
 #if RAW_BUFFER_LENGTH <= 254        // saves around 75 bytes program memory and speeds up ISR
     uint_fast8_t i;
@@ -1524,7 +1561,7 @@ void IRrecv::compensateAndPrintIRResultAsCArray(Print *aSerial, bool aOutputMicr
         aSerial->print(F("uint8_t rawTicks["));          // variable type, array name
     }
 
-    aSerial->print(decodedIRData.rawDataPtr->rawlen - 1, DEC);    // array size
+    aSerial->print(decodedIRData.rawlen - 1, DEC);    // array size
     aSerial->print(F("] = {"));    // Start declaration
 
 // Dump data
@@ -1533,7 +1570,7 @@ void IRrecv::compensateAndPrintIRResultAsCArray(Print *aSerial, bool aOutputMicr
 #else
     unsigned int i;
 #endif
-    for (i = 1; i < decodedIRData.rawDataPtr->rawlen; i++) {
+    for (i = 1; i < decodedIRData.rawlen; i++) {
         uint32_t tDuration = decodedIRData.rawDataPtr->rawbuf[i] * MICROS_PER_TICK;
 
         if (i & 1) {
@@ -1553,10 +1590,8 @@ void IRrecv::compensateAndPrintIRResultAsCArray(Print *aSerial, bool aOutputMicr
             tTicks = (tTicks > UINT8_MAX) ? UINT8_MAX : tTicks;
             aSerial->print(tTicks);
         }
-        if (i + 1 < decodedIRData.rawDataPtr->rawlen)
-            aSerial->print(',');                // ',' not required on last one
-        if (!(i & 1))
-            aSerial->print(' ');
+        if (i + 1 < decodedIRData.rawlen) aSerial->print(',');                // ',' not required on last one
+        if (!(i & 1)) aSerial->print(' ');
     }
 
 // End declaration
@@ -1585,9 +1620,9 @@ void IRrecv::compensateAndStoreIRResultInArray(uint8_t *aArrayPtr) {
 #if RAW_BUFFER_LENGTH <= 254        // saves around 75 bytes program memory and speeds up ISR
     uint_fast8_t i;
 #else
-        unsigned int i;
+    unsigned int i;
 #endif
-    for (i = 1; i < decodedIRData.rawDataPtr->rawlen; i++) {
+    for (i = 1; i < decodedIRData.rawlen; i++) {
         uint32_t tDuration = decodedIRData.rawDataPtr->rawbuf[i] * MICROS_PER_TICK;
         if (i & 1) {
             // Mark