+++ src/ir_FAST.hpp
@@ -96,13 +96,13 @@ void IRsend::sendFAST(uint8_t aCommand, int_fast8_t aNumberOfRepeats) {
 
 bool IRrecv::decodeFAST() {
 
-//    uint_fast8_t tRawlen = decodedIRData.rawDataPtr->rawlen; // Using a local variable does not improve code size
+//    uint_fast8_t tRawlen = decodedIRData.rawlen; // Using a local variable does not improve code size
 
     // Check we have the right amount of data (36). The +4 is for initial gap, start bit mark and space + stop bit mark.
-    if (decodedIRData.rawDataPtr->rawlen != ((2 * FAST_BITS) + 4)) {
+    if (decodedIRData.rawlen != ((2 * FAST_BITS) + 4)) {
         IR_DEBUG_PRINT(F("FAST: "));
         IR_DEBUG_PRINT(F("Data length="));
-        IR_DEBUG_PRINT(decodedIRData.rawDataPtr->rawlen);
+        IR_DEBUG_PRINT(decodedIRData.rawlen);
         IR_DEBUG_PRINTLN(F(" is not 36"));
         return false;
     }