+++ src/ir_Pronto.hpp
@@ -256,10 +256,10 @@ void IRrecv::compensateAndPrintIRResultAsPronto(Print *aSerial, uint16_t aFreque
     aSerial->print(F("char prontoData[] = \""));
     dumpNumber(aSerial, aFrequencyHertz > 0 ? learnedToken : learnedNonModulatedToken);
     dumpNumber(aSerial, toFrequencyCode(aFrequencyHertz));
-    dumpNumber(aSerial, (decodedIRData.rawDataPtr->rawlen + 1) / 2);
+    dumpNumber(aSerial, (decodedIRData.rawlen + 1) / 2);
     dumpNumber(aSerial, 0);
     uint16_t timebase = toTimebase(aFrequencyHertz);
-    compensateAndDumpSequence(aSerial, &decodedIRData.rawDataPtr->rawbuf[1], decodedIRData.rawDataPtr->rawlen - 1, timebase); // skip leading space
+    compensateAndDumpSequence(aSerial, &decodedIRData.rawDataPtr->rawbuf[1], decodedIRData.rawlen - 1, timebase); // skip leading space
     aSerial->println("\";");
 }
 
@@ -326,9 +326,9 @@ size_t IRrecv::compensateAndStorePronto(String *aString, uint16_t frequency) {
 
     size += dumpNumber(aString, frequency > 0 ? learnedToken : learnedNonModulatedToken);
     size += dumpNumber(aString, toFrequencyCode(frequency));
-    size += dumpNumber(aString, (decodedIRData.rawDataPtr->rawlen + 1) / 2);
+    size += dumpNumber(aString, (decodedIRData.rawlen + 1) / 2);
     size += dumpNumber(aString, 0);
-    size += compensateAndDumpSequence(aString, &decodedIRData.rawDataPtr->rawbuf[1], decodedIRData.rawDataPtr->rawlen - 1,
+    size += compensateAndDumpSequence(aString, &decodedIRData.rawDataPtr->rawbuf[1], decodedIRData.rawlen - 1,
             timebase); // skip leading space
 
     return size;