+++ src/ir_Sony.hpp
@@ -113,16 +113,16 @@ bool IRrecv::decodeSony() {
     }
 
     // Check we have enough data. +2 for initial gap and start bit mark and space minus the last/MSB space. NO stop bit! 26, 32, 42
-    if (decodedIRData.rawDataPtr->rawlen != (2 * SONY_BITS_MIN) + 2 && decodedIRData.rawDataPtr->rawlen != (2 * SONY_BITS_MAX) + 2
-            && decodedIRData.rawDataPtr->rawlen != (2 * SONY_BITS_15) + 2) {
+    if (decodedIRData.rawlen != (2 * SONY_BITS_MIN) + 2 && decodedIRData.rawlen != (2 * SONY_BITS_MAX) + 2
+            && decodedIRData.rawlen != (2 * SONY_BITS_15) + 2) {
         IR_DEBUG_PRINT(F("Sony: "));
         IR_DEBUG_PRINT(F("Data length="));
-        IR_DEBUG_PRINT(decodedIRData.rawDataPtr->rawlen);
+        IR_DEBUG_PRINT(decodedIRData.rawlen);
         IR_DEBUG_PRINTLN(F(" is not 12, 15 or 20"));
         return false;
     }
 
-    if (!decodePulseDistanceWidthData(&SonyProtocolConstants, (decodedIRData.rawDataPtr->rawlen - 1) / 2, 3)) {
+    if (!decodePulseDistanceWidthData(&SonyProtocolConstants, (decodedIRData.rawlen - 1) / 2, 3)) {
 #if defined(LOCAL_DEBUG)
         Serial.print(F("Sony: "));
         Serial.println(F("Decode failed"));
@@ -134,7 +134,7 @@ bool IRrecv::decodeSony() {
 //    decodedIRData.flags = IRDATA_FLAGS_IS_LSB_FIRST; // Not required, since this is the start value
     decodedIRData.command = decodedIRData.decodedRawData & 0x7F;  // first 7 bits
     decodedIRData.address = decodedIRData.decodedRawData >> 7;    // next 5 or 8 or 13 bits
-    decodedIRData.numberOfBits = (decodedIRData.rawDataPtr->rawlen - 1) / 2;
+    decodedIRData.numberOfBits = (decodedIRData.rawlen - 1) / 2;
     decodedIRData.protocol = SONY;
 
     //Check for repeat